+++ Source/DotNetWorkQueue/Queue/QueueWaitNoOp.cs
@@ -17,6 +17,7 @@
 //Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301  USA
 // ---------------------------------------------------------------------
 using System;
+using System.Threading.Tasks;
 
 namespace DotNetWorkQueue.Queue
 {
@@ -47,5 +48,9 @@ namespace DotNetWorkQueue.Queue
         {
 
         }
+        /// <summary>
+        /// Returns immediately; there is nothing to wait for.
+        /// </summary>
+        public ValueTask WaitAsync() => default;
     }
 }